chore: rename useRevokeBatch2
