+++ lib/hooks/ethereum/useRevokeBatch.tsx
